Extracting the business logic from the PickListComponent into the Container
